Flatten the transaction export styles
