+++ src/components/scenes/TransactionsExportScene.js
@@ -2,14 +2,15 @@
 
 import type { EdgeCurrencyWallet, EdgeGetTransactionsOptions } from 'edge-core-js'
 import React, { Component } from 'react'
-import { Platform, View } from 'react-native'
+import { Platform, StyleSheet, View } from 'react-native'
 import RNFS from 'react-native-fs'
 import Mailer from 'react-native-mail'
 import Share from 'react-native-share'
 
 import s from '../../locales/strings'
 import { PrimaryButton } from '../../modules/UI/components/Buttons/PrimaryButton.ui.js'
-import { styles } from '../../styles/scenes/TransactionsExportSceneStyle.js'
+import { THEME } from '../../theme/variables/airbitz.js'
+import { scale } from '../../util/scaling.js'
 import { sanitizeForFilename } from '../../util/utils.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 
@@ -149,3 +150,15 @@ export class TransactionsExportSceneComponent extends Component<Props> {
     )
   }
 }
+
+const rawStyles = {
+  actionButtonContainer: {
+    alignSelf: 'center',
+    width: '90%',
+    height: scale(THEME.BUTTONS.HEIGHT)
+  },
+  shim: {
+    height: scale(20)
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
